Move settings to core
