+++ app/core/database.py
@@ -1,7 +1,7 @@
 from typing import Annotated
 from fastapi import Depends
 from sqlmodel import Session, create_engine
-from app.settings import settings
+from app.core.settings import settings
 import app.models
 
 engine = create_engine(str(settings.postgres_url), echo=True)
